Move IProvideClassInfo to Interop OleAut32
