+++ src/System.Windows.Forms.Primitives/src/Interop/OleAut32/Interop.IProvideClassInfo.cs
@@ -1,0 +1,20 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Runtime.InteropServices;
+
+internal static partial class Interop
+{
+    internal static partial class Oleaut32
+    {
+        [ComImport]
+        [Guid("B196B283-BAB4-101A-B69C-00AA00341D07")]
+        [InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
+        public interface IProvideClassInfo
+        {
+            [PreserveSig]
+            HRESULT GetClassInfo(out Oleaut32.ITypeInfo ppTI);
+        }
+    }
+}
